// src/components/types.ts
export type Step =
  | "start"
  | "question_1"
  | "question_2"
  | "question_3"
  | "question_4"
  | "question_5"
  | "question_6"
  | "question_7"
  | "question_8"
  | "result_1"
  | "result_2"
  | "result_3"
  | "result_4"
  | "result_5"

export type ResultType = {
  result_1: number;
  result_2: number;
  result_3: number;
  result_4: number;
  result_5: number;
};